pdf-reader-ai/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pymupdf==1.23.7
chromadb==0.4.22
sentence-transformers==2.2.2
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.2
numpy==1.26.2
